# examples/v0.4.0-codegen-user-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User API",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "user"
    }
  ],
  "paths": {
    "/api/user/ping": {
      "get": {
        "operationId": "ping",
        "tags": [
          "user"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/user/users": {
      "get": {
        "operationId": "listUsers",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPageResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed"
          }
        }
      },
      "post": {
        "operationId": "createUser",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/user/users/{id}": {
      "get": {
        "operationId": "getUser",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed"
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUserReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "age": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UpdateUserReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ListUsersReq": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "pageSize"
        ]
      },
      "GetUserReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
      "DeleteUserReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
      "UserResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "UserPageResp": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page"
        ]
      },
      "OkResp": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "string"
          }
        },
        "required": [
          "ok"
        ]
      }
    }
  }
}