# examples/v0.4.0-codegen-user-api/src/dto/create-user-req.dto.ts
// Code generated by nofaultctl from api/user.api.ts. DO NOT EDIT.
import { IsEmail, IsInt, IsNotEmpty, IsOptional, IsString, Max, MaxLength, Min, MinLength } from '@nofault/rest';
export class CreateUserReq {
  @IsString()
  @IsNotEmpty()
  @MinLength(2)
  @MaxLength(32)
  name!: string;
  @IsString()
  @IsEmail()
  email!: string;
  @IsOptional()
  @IsInt()
  @Min(0)
  @Max(150)
  age!: number;
}