# examples/v0.4.0-codegen-user-api/src/dto/user-resp.dto.ts
// Code generated by nofaultctl from api/user.api.ts. DO NOT EDIT.
import { IsInt, IsString } from '@nofault/rest';
export class UserResp {
  @IsInt()
  id!: number;
  @IsString()
  name!: string;
  @IsString()
  email!: string;
}