# examples/v0.4.0-codegen-user-api/src/user/user.controller.ts
// Code generated by nofaultctl from api/user.api.ts. DO NOT EDIT.
import { UserService } from './user.service';
import { CreateUserReq } from '../dto/create-user-req.dto';
import { ListUsersReq } from '../dto/list-users-req.dto';
import { OkResp } from '../dto/ok-resp.dto';
import { UserPageResp } from '../dto/user-page-resp.dto';
import { UserResp } from '../dto/user-resp.dto';
import { Body, Controller, Delete, Get, Param, Post, Query, Validate } from '@nofault/rest';
@Controller({ path: '/api/user', middleware: ['RequestLogger'] })
export class UserController {
  constructor(private readonly service: UserService) {}

  @Get('/ping')
  async ping(): Promise<OkResp> {
    return this.service.ping();
  }

  @Validate(ListUsersReq)
  @Get('/users')
  async listUsers(@Query() listUsersReq: ListUsersReq): Promise<UserPageResp> {
    return this.service.listUsers(listUsersReq);
  }

  @Get('/users/:id')
  async getUser(@Param('id') id: number): Promise<UserResp> {
    return this.service.getUser({ id });
  }

  @Validate(CreateUserReq)
  @Post('/users')
  async createUser(@Body() createUserReq: CreateUserReq): Promise<UserResp> {
    return this.service.createUser(createUserReq);
  }

  @Delete('/users/:id')
  async deleteUser(@Param('id') id: number): Promise<OkResp> {
    return this.service.deleteUser({ id });
  }

}